+++ pine_cubit/__brick__/lib/cubits/{{file_name}}/{{file_name}}_cubit.dart
@@ -1,6 +1,7 @@
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:freezed_annotation/freezed_annotation.dart';
 import 'dart:async';
+{{#context}}import 'package:flutter/material.dart';{{/context}}
 
 part '{{file_name}}_state.dart';
 
@@ -8,7 +9,6 @@ part '{{file_name}}_cubit.freezed.dart';
 
 /// The {{cubit_name}}
 class {{cubit_name}} extends Cubit<{{state}}> {
-
   /// Create a new instance of [{{cubit_name}}].
   {{cubit_name}}() : super(const {{default_state}}());
   {{#methods}}
@@ -18,3 +18,8 @@ class {{cubit_name}} extends Cubit<{{state}}> {
   }
   {{/methods}}
 }
+
+{{#context}}extension {{cubit_name}}Extension on BuildContext {
+  /// Extension method used to get the [{{cubit_name}}] instance
+  {{cubit_name}} get {{cubit_name.camelCase()}} => read<{{cubit_name}}>();
+}{{/context}}