+++ pine_repository/__brick__/test/repositories/{{file_name}}/{{file_name}}_repository_test.dart
@@ -9,5 +9,11 @@ void main() {
     repository = const {{name}}RepositoryImpl();
   });
 
-  //TODO: Test your methods
+  {{#functions}}
+  /// Testing the method [{{method_name}}]
+  group('when the method {{method_name}} is called', () {
+
+  });
+
+  {{/functions}}
 }