+++ pine_retrofit/__brick__/lib/services/network/{{file_name}}/{{file_name}}_service.dart
@@ -1,5 +1,7 @@
 import 'package:dio/dio.dart';
 import 'package:retrofit/retrofit.dart';
+{{#context}}import 'package:flutter/material.dart';
+import 'package:provider/provider.dart';{{/context}}
 
 part '{{file_name}}_service.g.dart';
 
@@ -10,3 +12,8 @@ abstract class {{name}}Service {
 
   //TODO: Add your methods
 }
+
+{{#context}}extension {{name}}ServiceExtension on BuildContext {
+  /// Extension method used to get the [{{name}}Service] instance
+  {{name}}Service get {{name.camelCase()}}Repository => read<{{name}}Service>();
+}{{/context}}