+++ pine_service/__brick__/test/services/{{file_name}}/{{file_name}}_service_test.dart
@@ -9,5 +9,11 @@ void main() {
     service = const {{name}}ServiceImpl();
   });
 
-  //TODO: Test your methods
+  {{#functions}}
+  /// Testing the method [{{method_name}}]
+  group('when the method {{method_name}} is called', () {
+
+  });
+
+  {{/functions}}
 }