+++ pine_repository/CHANGELOG.md
@@ -1,3 +1,8 @@
+# 0.2.0
+
+Added comma-separated methods generation
+Added repository extension on context
+
 # 0.1.0
 
 First release