+++ pine_retrofit/CHANGELOG.md
@@ -1,3 +1,7 @@
+# 0.2.0
+
+Added service extension on context
+
 # 0.1.0
 
 First release